requirements.txt
beautifulsoup4==4.6.0
requests==2.21.0
docopt==0.6.2
pandas==0.24.1
backoff==1.8.0
progressbar2==3.39.2
selenium==3.141.0
